+++ src/test-utils.ts
@@ -38,8 +38,6 @@ export async function expectErrorResponse(
   return data;
 }
 
-// API test pattern helper
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function runApiTests<T extends (...args: any[]) => Promise<any>>(
   name: string,
   fn: T,